+++ cmd/neofs-adm/internal/modules/fschain/container.go
@@ -58,7 +58,7 @@ func dumpContainers(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return fmt.Errorf("can't create N3 client: %w", err)
 	}
@@ -118,7 +118,7 @@ func dumpContainers(cmd *cobra.Command, _ []string) error {
 }
 
 func listContainers(cmd *cobra.Command, _ []string) error {
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return fmt.Errorf("can't create N3 client: %w", err)
 	}