+++ cmd/neofs-adm/internal/modules/fschain/load.go
@@ -38,7 +38,7 @@ func reportsFunc(cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("invalid container ID: %w", err)
 	}
 
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return fmt.Errorf("can't create N3 client: %w", err)
 	}
@@ -122,7 +122,7 @@ func loadSummaryFunc(cmd *cobra.Command, args []string) error {
 		}
 	}
 
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return fmt.Errorf("can't create N3 client: %w", err)
 	}