+++ cmd/neofs-adm/internal/modules/fschain/nodes.go
@@ -40,7 +40,7 @@ func nodesFunc(cmd *cobra.Command, _ []string) error {
 		return fmt.Errorf("invalid container ID: %w", err)
 	}
 
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return fmt.Errorf("can't create N3 client: %w", err)
 	}