+++ cmd/neofs-adm/internal/modules/fschain/notary.go
@@ -79,7 +79,7 @@ func depositNotary(cmd *cobra.Command, _ []string) error {
 		}
 	}
 
-	c, err := getN3Client(viper.GetViper())
+	c, err := GetN3Client(viper.GetViper())
 	if err != nil {
 		return err
 	}